Remove validation of OCP selectors
