+++ config/webhook/manifests.yaml
@@ -39,27 +39,6 @@ webhooks:
       service:
         name: webhook-service
         namespace: system
-        path: /validate-namespace-node-selector
-    failurePolicy: Fail
-    matchPolicy: Equivalent
-    name: validate-namespace-node-selector.appuio.io
-    rules:
-      - apiGroups:
-          - ""
-        apiVersions:
-          - v1
-        operations:
-          - CREATE
-          - UPDATE
-        resources:
-          - namespaces
-    sideEffects: None
-  - admissionReviewVersions:
-      - v1
-    clientConfig:
-      service:
-        name: webhook-service
-        namespace: system
         path: /validate-request-ratio
     failurePolicy: Ignore
     matchPolicy: Equivalent
@@ -75,30 +54,4 @@ webhooks:
         resources:
           - '*'
         scope: Namespaced
-    sideEffects: None
-  - admissionReviewVersions:
-      - v1
-    clientConfig:
-      service:
-        name: webhook-service
-        namespace: system
-        path: /validate-workload-node-selector
-    failurePolicy: Fail
-    matchPolicy: Equivalent
-    name: validate-workload-node-selector.appuio.io
-    rules:
-      - apiGroups:
-          - '*'
-        apiVersions:
-          - v1
-        operations:
-          - CREATE
-          - UPDATE
-        resources:
-          - pods
-          - daemonsets
-          - deployments
-          - jobs
-          - statefulsets
-          - cronjobs
     sideEffects: None
