+++ main.go
@@ -17,7 +17,6 @@ import (
 	"github.com/appuio/appuio-cloud-agent/controllers"
 	"github.com/appuio/appuio-cloud-agent/ratio"
 	"github.com/appuio/appuio-cloud-agent/skipper"
-	"github.com/appuio/appuio-cloud-agent/validate"
 	"github.com/appuio/appuio-cloud-agent/webhooks"
 )
 
@@ -108,27 +107,6 @@ func main() {
 }
 
 func registerNodeSelectorValidationWebhooks(mgr ctrl.Manager, skip skipper.Skipper, conf Config) {
-	ans := &validate.AllowedLabels{}
-	for k, v := range conf.AllowedNodeSelectors {
-		if err := ans.Add(k, v); err != nil {
-			setupLog.Error(err, "unable to add allowed node selector")
-			os.Exit(1)
-		}
-	}
-
-	mgr.GetWebhookServer().Register("/validate-namespace-node-selector", &webhook.Admission{
-		Handler: &webhooks.NamespaceNodeSelectorValidator{
-			Skipper:               skip,
-			AllowedNodeSelectors:  ans,
-			DenyEmptyNodeSelector: conf.NamespaceDenyEmptyNodeSelector,
-		},
-	})
-	mgr.GetWebhookServer().Register("/validate-workload-node-selector", &webhook.Admission{
-		Handler: &webhooks.WorkloadNodeSelectorValidator{
-			Skipper:              skip,
-			AllowedNodeSelectors: ans,
-		},
-	})
 	mgr.GetWebhookServer().Register("/mutate-pod-node-selector", &webhook.Admission{
 		Handler: &webhooks.PodNodeSelectorMutator{
 			Skipper: &skipper.NonOrganizationNamespaceSkipper{
