+++ webhooks/pod_node_selector_mutator_test.go
@@ -7,10 +7,14 @@ import (
 	"github.com/stretchr/testify/require"
 	"gomodules.xyz/jsonpatch/v2"
 	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
+	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
+	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
+	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 
 	"github.com/appuio/appuio-cloud-agent/skipper"
 )
@@ -102,5 +106,47 @@ func Test_PodNodeSelectorMutator_Handle(t *testing.T) {
 			require.ElementsMatch(t, tc.patch, resp.Patches)
 			require.Equal(t, tc.allowed, resp.Allowed)
 		})
+	}
+}
+
+func newNamespace(name string, labels, annotations map[string]string) *corev1.Namespace {
+	return &corev1.Namespace{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "Namespace",
+			APIVersion: "v1",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:        name,
+			Labels:      labels,
+			Annotations: annotations,
+		},
+	}
+}
+
+func decoder(t *testing.T) *admission.Decoder {
+	t.Helper()
+
+	scheme := runtime.NewScheme()
+	utilruntime.Must(clientgoscheme.AddToScheme(scheme))
+
+	decoder, err := admission.NewDecoder(scheme)
+	require.NoError(t, err)
+
+	return decoder
+}
+
+func newPod(namespace, name string, nodeSelector map[string]string) *corev1.Pod {
+	return &corev1.Pod{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "Pod",
+			APIVersion: "v1",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      name,
+			Namespace: namespace,
+		},
+		Spec: corev1.PodSpec{
+			NodeSelector: nodeSelector,
+		},
 	}
 }
